added sqrl algorithm

+++ src/main/java/me/ilizin/spring_demo/springboot_demo/api_rest_demo/controllers/SqrtRestController.java
@@ -15,7 +15,7 @@ public class SqrtRestController {
     }
 
     @GetMapping("/sqrt/{value}")
-    public int sqrt(@PathVariable int value) {
+    public double sqrt(@PathVariable int value) {
         return sqrtService.sqrt(value);
     }
 }
