+++ src/main/java/me/ilizin/spring_demo/springboot_demo/api_rest_demo/services/interfaces/SqrtService.java
@@ -2,5 +2,5 @@ package me.ilizin.spring_demo.springboot_demo.api_rest_demo.services.interfaces;
 
 public interface SqrtService {
 
-    int sqrt(int value);
+    double sqrt(int value);
 }
